feat: refactor code a bit
